==> notebooks/0_setup.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

This notebook is just designed to check you have all the required packages installed. If any line fails due to the package note being available, uncomment the lines at the top of the cell and re-run to install the packages required

```{r}
#install.packages(c("Hmisc", "tidyverse"))
library(Hmisc)
library(tidyverse)
```

```{r}
#if (!requireNamespace("BiocManager", quietly = TRUE))
#    install.packages("BiocManager")

#BiocManager::install(c("biobroom", "MSnbase", "goseq", "GO.db", "UniProt.ws", "limma"))

library(limma)
library(biobroom)
library(MSnbase)
library(goseq)
library(GO.db)
library(UniProt.ws)
```
